+++ src/main/java/com/pepper/project/pm/guide/domain/GuidePm.java
@@ -25,7 +25,7 @@ public class GuidePm extends BaseEntity{
     private String content;
 
     /** 所属物业 */
-    private Integer propertyId;
+    private Long propertyId;
 
     /** 区域状态（0正常 1关闭） */
     private String status;
@@ -72,11 +72,11 @@ public class GuidePm extends BaseEntity{
         this.content = content;
     }
 
-    public Integer getPropertyId() {
+    public Long getPropertyId() {
         return propertyId;
     }
 
-    public void setPropertyId(Integer propertyId) {
+    public void setPropertyId(Long propertyId) {
         this.propertyId = propertyId;
     }
